1st Math lab
Lab in which I had to take in a number, find the square root of it, set it to a certain power, find the absolute value of it determine if it is true or false and square the number. Then, I had to take in a list of doubles and find the sum of them, the difference of them, the product of them and then the quotient of them.